// src/widgets/Item/models/types.ts
export type ProductItemProps = {
  image: string;
  background: string;
  itemName: string;
  price: number;
  author: string;
  category: string;
  isFavorite: boolean;
  profile: string;
  id: string;
  time: string;
};

export type ItemProps = {
  image: string;
  background: string;
  itemName: string;
  price: number;
  author: string;
  category: string;
  isFavorite: boolean;
  profile: string;
  id: string;
};